+++ Civi/DataProcessor/FilterHandler/AbstractFieldFilterHandler.php
@@ -6,13 +6,10 @@
 
 namespace Civi\DataProcessor\FilterHandler;
 
-use Civi\DataProcessor\DataFlow\CombinedDataFlow\CombinedSqlDataFlow;
 use Civi\DataProcessor\DataFlow\InMemoryDataFlow;
 use Civi\DataProcessor\DataFlow\SqlDataFlow;
-use Civi\DataProcessor\DataFlow\SqlTableDataFlow;
 use Civi\DataProcessor\Exception\DataSourceNotFoundException;
 use Civi\DataProcessor\Exception\FieldNotFoundException;
-use Civi\DataProcessor\Source\AbstractCivicrmEntitySource;
 use CRM_Dataprocessor_ExtensionUtil as E;
 
 abstract class AbstractFieldFilterHandler extends AbstractFilterHandler {
@@ -120,16 +117,6 @@ abstract class AbstractFieldFilterHandler extends AbstractFilterHandler {
       $this->filterClass = new InMemoryDataFlow\SimpleFilter($this->inputFieldSpecification->getName(), $filter['op'], $filter['value']);
       $dataFlow->addFilter($this->filterClass);
     }
-  }
-
-  protected function getTableAlias(SqlDataFlow $dataFlow) {
-    $tableAlias = $dataFlow->getName();
-    if ($dataFlow instanceof SqlTableDataFlow) {
-      $tableAlias = $dataFlow->getTableAlias();
-    } elseif ($dataFlow instanceof CombinedSqlDataFlow) {
-      $tableAlias = $dataFlow->getPrimaryTableAlias();
-    }
-    return $tableAlias;
   }
 
 }